requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
datasets==4.1.1
dill==0.4.0
filelock==3.19.1
fonttools==4.60.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
multitasking==0.0.12
networkx==3.3
nltk==3.9.1
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.0.0
platformdirs==4.4.0
praw==7.8.1
prawcore==2.4.0
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.8
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.56.2
tweepy==4.16.0
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.66
